+++ src/main/java/com/dlqudtjs/codingbattle/service/session/SessionService.java
@@ -2,9 +2,9 @@ package com.dlqudtjs.codingbattle.service.session;
 
 public interface SessionService {
 
-    void enterRoom(String userId, Integer roomId);
+    void enterRoom(String userId, Long roomId);
 
     void leaveRoom(String userId);
 
-    Integer getUserInRoomId(String userId);
+    Long getUserInRoomId(String userId);
 }
